refactor: modelをdomain_modelに集約
