+++ packages/domain_model/lib/internal_domain_model.dart
@@ -4,9 +4,5 @@
 library;
 
 export 'src/force_update_settings_state.dart';
-export 'src/force_update_target_version.dart';
-export 'src/internal_domain_model_base.dart';
-export 'src/maintenance_mode_settings_state.dart';
-export 'src/version_string.dart';
-
-// TODO: Export any libraries intended for clients of this package.
+export 'src/operating_system.dart';
+export 'src/operational_settings.dart';
