--- packages/domain_model/lib/src/maintenance_mode_settings_state.dart
@@ -1,11 +1,0 @@
-import 'package:freezed_annotation/freezed_annotation.dart';
-
-part 'maintenance_mode_settings_state.freezed.dart';
-
-@freezed
-abstract class MaintenanceModeSettingsState
-    with _$MaintenanceModeSettingsState {
-  const factory MaintenanceModeSettingsState({
-    @Default(false) bool enabled,
-  }) = _MaintenanceModeSettingsState;
-}
