+++ packages/domain_model/lib/src/operating_system.dart
