+++ packages/domain_model/lib/src/operational_settings.dart
@@ -1,5 +1,5 @@
-import 'package:features_force_update/src/domain/models/operating_system.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
+import 'package:internal_domain_model/src/operating_system.dart';
 import 'package:pub_semver/pub_semver.dart';
 
 part 'operational_settings.freezed.dart';
