+++ packages/domain_model/lib/src/operational_settings.freezed.dart
