+++ packages/domain_model/pubspec.yaml
@@ -12,6 +12,7 @@ dependencies:
   flutter:
     sdk: flutter
   freezed_annotation: ^3.0.0
+  pub_semver: ^2.2.0
 
 dev_dependencies:
   build_runner: ^2.4.15
