--- packages/domain_model/test/internal_domain_model_test.dart
@@ -1,16 +1,0 @@
-import 'package:internal_domain_model/internal_domain_model.dart';
-import 'package:test/test.dart';
-
-void main() {
-  group('A group of tests', () {
-    final awesome = Awesome();
-
-    setUp(() {
-      // Additional setup goes here.
-    });
-
-    test('First Test', () {
-      expect(awesome.isAwesome, isTrue);
-    });
-  });
-}
