+++ packages/domain_model/test/operational_settings_test.dart
@@ -1,7 +1,6 @@
-import 'package:features_force_update/src/domain/models/operating_system.dart';
-import 'package:features_force_update/src/domain/models/operational_settings.dart';
-import 'package:flutter_test/flutter_test.dart';
+import 'package:internal_domain_model/internal_domain_model.dart';
 import 'package:pub_semver/pub_semver.dart';
+import 'package:test/test.dart';
 
 void main() {
   group('OperationalSettings', () {
