+++ packages/features/force_update/lib/force_update.dart
@@ -1,5 +1,3 @@
 export 'package:pub_semver/pub_semver.dart';
 
-export 'src/domain/models/force_update_settings_state.dart';
-export 'src/domain/models/operational_settings.dart';
 export 'src/provider/force_update_provider.dart';
