+++ packages/features/force_update/lib/src/data/repositories/operational_setting_repository.dart
@@ -1,4 +1,4 @@
-import 'package:features_force_update/src/domain/models/operational_settings.dart';
+import 'package:internal_domain_model/internal_domain_model.dart';
 import 'package:pub_semver/pub_semver.dart';
 
 class OperationalSettingRepository {
