+++ packages/features/force_update/lib/src/use_cases/check_force_update_use_case.dart
@@ -1,6 +1,5 @@
 import 'package:features_force_update/src/data/repositories/operational_setting_repository.dart';
-import 'package:features_force_update/src/domain/models/operating_system.dart';
-import 'package:features_force_update/src/domain/models/operational_settings.dart';
+import 'package:internal_domain_model/internal_domain_model.dart';
 import 'package:pub_semver/pub_semver.dart';
 
 class CheckForceUpdateUseCase {
