+++ packages/features/force_update/pubspec.yaml
@@ -16,6 +16,8 @@ dependencies:
     path: ../../cores/model
   cores_ui:
     path: ../../cores/ui
+  internal_domain_model:
+    path: ../../domain_model
   flutter:
     sdk: flutter
   flutter_riverpod: ^2.6.1
